requirements.txt
streamlit
transformers
chromadb
nltk
pandas
qdrant-client
python-dotenv
torch==2.2.2
torchvision==0.17.2
transformers==4.40.2
sentence-transformers==2.7.0